requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==3.0.5
channels-redis==4.2.0
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
Django==4.2.13
django-allauth==0.63.3
django-crispy-forms==2.1
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
redis==5.0.5
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
zope.interface==6.4.post2
